backend/requirements.txt
attrs==19.3.0
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
convertdate==2.2.0
-e git+https://github.com/sandeshbhatjr/energy-prediction.git@960b82abfe8a394cfc46134c3702f0dccc43a2af#egg=empredictor&subdirectory=backend
entsoe-py==0.2.11
Flask==1.1.1
Flask-Cors==3.0.8
holidays==0.10.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jplephem==2.13
lunardate==0.2.0
MarkupSafe==1.1.1
more-itertools==8.2.0
numexpr==2.7.1
numpy==1.18.1
packaging==20.3
pandas==1.0.1
patsy==0.5.1
pluggy==0.13.1
py==1.8.1
pyCalverter==1.6.1
pyluach==1.0.1
PyMeeus==0.3.7
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scipy==1.4.1
sgp4==2.4
six==1.14.0
skyfield==1.17
skyfield-data==0.1.0
soupsieve==2.0
statsmodels==0.11.1
tables==3.6.1
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.0
zipp==3.1.0
